requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-smtp-ssl==1.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==21.2.0
idna==3.4
mailgun==0.1.1
marshmallow==3.20.1
mysqlclient==2.2.1
oauthlib==3.2.2
packaging==23.2
pyactiveresource==1.0.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
